rename : 리사이클 관련 파일 이름 수정
- 리사이클러뷰 관련 파일의 이름을 수정합니다
- 다른 뷰에서도 겹치는 부분이 많기에 재활용을 위해 이름을 수정합니다
@Hah_nna
