+++ app/src/main/java/com/jeong/sesac/sai/recycler/gridRecycler/DiffUtilGridNotes.kt
@@ -1,9 +1,9 @@
-package com.jeong.sesac.sai.recycler.recentlyFoundNotesRecycler
+package com.jeong.sesac.sai.recycler.gridRecycler
 
 import androidx.recyclerview.widget.DiffUtil
 import com.jeong.sesac.sai.util.WeeklyNotesInfo
 
-class DiffUtilRecentlyFoundNotes : DiffUtil.ItemCallback<WeeklyNotesInfo>() {
+class DiffUtilGridNotes : DiffUtil.ItemCallback<WeeklyNotesInfo>() {
     override fun areContentsTheSame(oldItem: WeeklyNotesInfo, newItem: WeeklyNotesInfo): Boolean {
         return oldItem.id == newItem.id
     }
