+++ app/src/main/java/com/jeong/sesac/sai/recycler/gridRecycler/GridNotesViewHolder.kt
@@ -1,0 +1,7 @@
+package com.jeong.sesac.sai.recycler.gridRecycler
+
+import androidx.recyclerview.widget.RecyclerView
+import com.jeong.sesac.sai.databinding.ItemRecentlyFoundNoteBinding
+
+class GridNotesViewHolder(val binding: ItemRecentlyFoundNoteBinding) : RecyclerView.ViewHolder(binding.root) {
+}
