+++ app/src/main/java/com/jeong/sesac/sai/recycler/horizontalRecycler/DiffUtilHorizontalNotes.kt
@@ -1,9 +1,9 @@
-package com.jeong.sesac.sai.recycler.weeklynotes
+package com.jeong.sesac.sai.recycler.horizontalRecycler
 
 import androidx.recyclerview.widget.DiffUtil
 import com.jeong.sesac.sai.util.WeeklyNotesInfo
 
-class DiffUtilWeeklyNotes : DiffUtil.ItemCallback<WeeklyNotesInfo>(){
+class DiffUtilHorizontalNotes : DiffUtil.ItemCallback<WeeklyNotesInfo>(){
 
     override fun areItemsTheSame(oldItem: WeeklyNotesInfo, newItem: WeeklyNotesInfo): Boolean {
         return oldItem.id == newItem.id
