+++ app/src/main/java/com/jeong/sesac/sai/recycler/horizontalRecycler/HorizontalNotesViewHolder.kt
@@ -1,0 +1,7 @@
+package com.jeong.sesac.sai.recycler.horizontalRecycler
+
+import androidx.recyclerview.widget.RecyclerView
+import com.jeong.sesac.sai.databinding.ItemWeeklyNotesBinding
+
+class HorizontalNotesViewHolder(val binding : ItemWeeklyNotesBinding) : RecyclerView.ViewHolder(binding.root) {
+}
